+++ src/php/Command/Repl/ColorStyle.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Commands\Repl;
+namespace Phel\Command\Repl;
 
 /** @psalm-immutable */
 final class ColorStyle
