+++ src/php/Command/Repl/ReplCommandSystemIo.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Commands\Repl;
+namespace Phel\Command\Repl;
 
 final class ReplCommandSystemIo implements ReplCommandIoInterface
 {
