+++ tests/php/Integration/Command/Fixtures/test-cmd-project-failure/composer.json
