+++ tests/php/Integration/Command/Fixtures/test-cmd-project-success/composer.json
