+++ tests/php/Integration/Command/Fixtures/test-script-without-ns.phel
