+++ tests/php/Integration/Command/Fixtures/test-script.phel
